requirements.txt
# SRP Acceleration Prediction - Project Dependencies
# Author: Divyanshu Panday
# Date: October 2025

# Data Acquisition & Processing
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file operations
python-dotenv>=1.0.0
numpy>=1.24.0
poliastro

# Orbital Mechanics & Astronomy
sgp4>=2.23
astropy>=5.3.0
skyfield>=1.46  # High-precision position calculation

# Data Analysis & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
plotly>=5.17.0  # Interactive 3D orbit visualization

# Machine Learning
scikit-learn>=1.3.0  # Random Forest, metrics, preprocessing
#xgboost>=2.0.0  # XGBoost regressor (optional)
joblib>=1.3.0  # Model serialization (.pkl files)

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional utilities
python-dateutil>=2.8.2
streamlit

lightgbm
scipy
joblib
python-dotenv
